去除vs编译警告（64bit）

+++ inc/simple/string.h
@@ -255,10 +255,10 @@ std::string			string_base64_encode(const unsigned char* s, unsigned long len);
 std::string			string_base64_decode(const unsigned char* s, unsigned long len);
 
 inline	std::string	string_base64_encode(const char* s, unsigned int len) {
-    return	string_base64_encode((const unsigned char*)s, len);
+    return	string_base64_encode((const unsigned char*)s, (unsigned long)len);
 }
 inline	std::string	string_base64_encode(const std::string& s) {
-    return	string_base64_encode((const unsigned char*)s.c_str(), s.size());
+    return	string_base64_encode((const unsigned char*)s.c_str(), (unsigned long)s.size());
 }
 
 inline  std::string	string_base64_decode(const std::string& s) {
@@ -272,10 +272,10 @@ std::string			string_qp_encode(const unsigned char* s, unsigned long len);
 std::string			string_qp_decode(const unsigned char* s, unsigned long len);
 
 inline	std::string	string_qp_encode(const char* s, unsigned int len) {
-    return	string_qp_encode((const unsigned char*)s, len);
+    return	string_qp_encode((const unsigned char*)s, (unsigned long)len);
 }
 inline	std::string	string_qp_encode(const std::string& s) {
-    return	string_qp_encode((const unsigned char*)s.c_str(), s.size());
+    return	string_qp_encode((const unsigned char*)s.c_str(), (unsigned long)s.size());
 }
 inline  std::string	string_qp_decode(const std::string& s) {
     return  string_qp_decode((const unsigned char*)s.c_str(), (unsigned long)s.size());
